Added long lost Algebraic Identities tests
